grid: Move browser not. toggle into notification tab, remove dnd
